requirements.txt
streamlit>=1.27.0
python-dotenv>=1.0.1
groq>=0.3.0
langchain-community>=0.0.73
PyPDF2==3.0.1
python-docx>=0.8.11
